+++ Squalr.Engine.Memory/IMemoryReader.cs
@@ -17,7 +17,7 @@
         /// <param name="address">The address to read from.</param>
         /// <param name="success">Whether or not the read succeeded.</param>
         /// <returns>The value read from memory.</returns>
-        Object Read(DataType elementType, IntPtr address, out Boolean success);
+        Object Read(DataType elementType, UInt64 address, out Boolean success);
 
         /// <summary>
         /// Reads a value from the opened processes memory.
@@ -26,7 +26,7 @@
         /// <param name="address">The address to read from.</param>
         /// <param name="success">Whether or not the read succeeded.</param>
         /// <returns>The value read from memory.</returns>
-        T Read<T>(IntPtr address, out Boolean success);
+        T Read<T>(UInt64 address, out Boolean success);
 
         /// <summary>
         /// Reads an array of bytes from the opened processes memory.
@@ -35,9 +35,9 @@
         /// <param name="count">The number of bytes to read.</param>
         /// <param name="success">Whether or not the read succeeded.</param>
         /// <returns>The array of bytes read from memory, if the read succeeded.</returns>
-        Byte[] ReadBytes(IntPtr address, Int32 count, out Boolean success);
+        Byte[] ReadBytes(UInt64 address, Int32 count, out Boolean success);
 
-        IntPtr EvaluatePointer(IntPtr address, IEnumerable<Int32> offsets);
+        UInt64 EvaluatePointer(UInt64 address, IEnumerable<Int32> offsets);
     }
     //// End interface
 }
