+++ Squalr.Engine.Memory/IMemoryWriter.cs
@@ -15,7 +15,7 @@
         /// <param name="elementType">The data type to write.</param>
         /// <param name="address">The address to write to.</param>
         /// <param name="value">The value to write.</param>
-        void Write(DataType elementType, IntPtr address, Object value);
+        void Write(DataType elementType, UInt64 address, Object value);
 
         /// <summary>
         /// Writes a value to memory in the opened process.
@@ -23,14 +23,14 @@
         /// <typeparam name="T">The data type to write.</typeparam>
         /// <param name="address">The address to write to.</param>
         /// <param name="value">The value to write.</param>
-        void Write<T>(IntPtr address, T value);
+        void Write<T>(UInt64 address, T value);
 
         /// <summary>
         /// Writes a value to memory in the opened process.
         /// </summary>
         /// <param name="address">The address to write to.</param>
         /// <param name="values">The value to write.</param>
-        void WriteBytes(IntPtr address, Byte[] values);
+        void WriteBytes(UInt64 address, Byte[] values);
     }
     //// End interface
 }
